✨ Refactor card service component: move files to components directory, update imports, and enhance styling and functionality
