+++ web/src/app/components/card-service/card-service.css
@@ -1,10 +1,9 @@
 
 .service-item {
-    height: 220px;
     border-radius: 10px;
     text-align: center;
     width: 300px;
-    height: 200px;
+    height: 220px;
     perspective: 1000px; /* Remove this if you don't want the 3D effect */
     margin-bottom: 40px;
 }
@@ -71,4 +70,13 @@
         color: var(--dark-gray-color);
         padding: 10px;
     }
+}
+
+.connect-button {
+    display: flex;
+    height: auto;
+    justify-content: end;
+    align-items: center;
+    padding: 15px;
+    padding-top: 0;
 }
