+++ web/src/app/pages/service-management-page/service-management-page.css
@@ -1,0 +1,18 @@
+.explorer-page {
+    text-align: center;
+    padding-top: 100px;
+    padding-left: 200px;
+    padding-right: 200px;
+    h1 {
+        font-size: 45px;
+        font-weight: bold;
+        color: var(--black-color);
+    }
+}
+
+.services-list {
+    margin-top: 50px;
+    display: grid;
+    grid-template-columns: repeat(5, 1fr);
+    gap: 40px;
+}
