+++ web/src/app/pages/service-management-page/service-management-page.html
@@ -1,1 +1,17 @@
 <app-area-header-dashboard [title]="'SERVICES'"></app-area-header-dashboard>
+<div class="explorer-page">
+    <h1>EXPLORE</h1>
+    <app-text-field-component
+        [label]="'Search service...'"
+        [type]="'text'"
+        [(value)]="searchService"
+        (valueChange)="searchService = $event"
+    ></app-text-field-component>
+    <div class="services-list">
+        <app-card-service *ngFor="let service of filteredSerice()"
+            [name]="service.name"
+            [description]="service.description"
+            [connectedShow]="true"
+        ></app-card-service>
+    </div>
+</div>
